Listen for/respond to loader/canvas signals
Listen for and respond to signals from loader and canvas slaves.

When an electrode is selected, send a request to the
`wheelerlab.electrode_controller_plugin` to toggle the state of the
electrode.

When a pair of electrodes is selected, find the shortest path between
the electrodes and send a request to the
`wheelerlab.droplet_planning_plugin` to add the route.

When the state of one or more electrodes has changed, update the states
in the canvas slave accordingly and render the new states to update the
device drawing.

When the routes have changed, update the routes in the canvas slave
accordingly and render the new routes to update the device drawing.