[NB] Replace DeviceLoader slave...
Replace `DeviceLoader` slave view with `DevicePluginConnection` slave
view.

The `DevicePluginConnection` slave view inherits from the generic
`PluginConnection` slave view, which provides a UI to create a plugin
instance and connect the plugin to a hub at the specified URL.

TODO
----

The `PluginConnection` slave view should be moved to either the
`pygtkhelpers` package or the `zmq-plugin` package, since its
functionality is not specific to the DMF device user interface.