[NB][FIX] Do not call refresh_gui on resize/move
Prior to this commit, the `pygtkhelpers.utils.refresh_gui` function was
called when the window was resized or moved, and in the heart beat timer
callback.

Upon investigation, it has been determined that calls to `refresh_gui` in
these cases lead to issue #3.

As of this commit, `refresh_gui` is no longer called on window resize/move
or in the heart-beat callback.

This commit fixes #3.