fix(routes): route continues after button release
Prior to this commit, the following setup:

 1. Press left mouse down over electrode in device view.
 2. Drag outside of electrodes area (while button is still pressed down).
 3. Release button.
 4. Drag mouse back over electrodes.

resulted in the route continuing as if the button was still pressed down
after re-entering electrodes area, even though button was not pressed.

In this commit, clear any temporary route when the mouse button is
released.  If no valid route was drawn (e.g., in the setup listed above),
the temporary route is cleared and no route is drawn.  As part of this
commit, add the `routes-set` GTK signal to the `DmfDeviceCanvas` canvas
slave, emitted whenever the `DmfDeviceCanvas.df_routes` routes table is
modified.

Fixes sci-bots/microdrop#256.