fix(plugin): queue redraw on alphas update
Prior to this commit, when the surface alphas were updated from the ZeroMQ
plugin interface, the UI did not reflect the changes until the next time
the canvas was redrawn.

In this commit, queue a redraw of the canvas after the surface alphas are
updated to reflect the changes immediately.