feat(commands): flatten electrode/route menus
Prior to this commit, commands under the "Electrodes" and "Routes" command
sub-menus in the context menu were further split into sub-menus grouped by
plugin name.  However, this resulted in a deep menu hierarchy that was
tedious to navigate.

In this commit, flatten the commands under the "Electrodes" and "Routes"
sub-menus in context menu, resulting in a context menu structure like:

 - <global command ...>
 - <global command ...>
 - ...
 - <separator>
 - "Electrodes"
   - <electrode command ...>
   - <electrode command ...>
   - ...
 - <separator>
 - "Routes"
   - <route command ...>
   - <route command ...>
   - ...